+++ tosixinch/process/sample.py
@@ -12,13 +12,8 @@ from copy import deepcopy
 import logging
 import re
 
-from tosixinch import _ImportError
+from tosixinch import lxml_html
 from tosixinch.clean import KEEP_STYLE  # noqa: F401
-
-try:
-    import lxml.html
-except ImportError:
-    lxml = _ImportError('lxml')
 
 logger = logging.getLogger(__name__)
 
@@ -30,11 +25,8 @@ self_closing_tags = (
     'source', 'track', 'wbr',
 )
 
-# These short forms are used only in ``process`` package.
-# 'tostring' is perhaps only in doctests.
-# Other modules use longer 'lxml.html...' form.
-fromstring = lambda el: lxml.html.fromstring(el)
-tostring = lambda el: lxml.html.tostring(el, encoding='unicode')
+fromstring = lambda el: lxml_html.fromstring(el)
+tostring = lambda el: lxml_html.tostring(el, encoding='unicode')
 
 
 # lxml.html module's parsing functions are
@@ -214,18 +206,6 @@ def get_metadata(el):
     return dict(authors=authors, description=description,
         generator=generator, keywords=keywords,
         created=created, modified=modified)
-
-
-def build_class_xpath(path):
-    """Create a selector for a class from multi classes element.
-
-    >>> build_class_xpath('//div[@class=="main-article"]')
-    '//div[contains(concat(" ", normalize-space(@class), " "), " main-article ")]'
-    """  # noqa: E501 line too long
-    pat = r'([a-zA-Z]+|[hH][1-6]|\*)\[@class==([\'"])([_a-zA-Z0-9-]+)\2\]'
-    pat = re.compile(pat)
-    repl = r'\1[contains(concat(" ", normalize-space(@class), " "), " \3 ")]'
-    return pat.sub(repl, path)
 
 
 # ----------------------------------------------------------
@@ -571,10 +551,10 @@ def reddit_indent(doc):
     They are in the site's external css.
     So we have to insert inline css made from it.
     """
-    for el in doc.xpath(build_class_xpath('//div[@class=="comment"]')):
+    for el in doc.xpath('//div[@class=="comment"]'):
         el.set('style', 'margin-left:8px;')
         el.classes.add(KEEP_STYLE)
-        path = build_class_xpath('.//p[@class="tagline"]/a[@class=="author"]')
+        path = './/p[@class="tagline"]/a[@class=="author"]'
         for e in el.xpath(path):
             e.set('style', 'font-weight:bold;')
             e.classes.add(KEEP_STYLE)