+++ tosixinch/tests/test_clean.py
@@ -3,27 +3,27 @@ import lxml.html
 import pytest
 
 from tosixinch import clean
-
-fromstring = lxml.html.fromstring
-tostring = lambda el: lxml.html.tostring(el, encoding='unicode')
+from tosixinch import lxml_html
 
 
-class TestFilteredIter:
+def test_conditioned_iter():
+    data = '<div><h3>aaa</h3><div><p>bbb<span>ccc</span></p></div></div>'
+    expected = [
+        '<div><h3>aaa</h3><div><p>bbb<span>ccc</span></p></div></div>',
+        '<h3>aaa</h3>',
+        '<div><p>bbb<span>ccc</span></p></div>']
 
-    def test(self):
-        data = '<div><h3>aaa</h3><div><p>bbb<span>ccc</span></p></div></div>'
-        expected = [
-            '<div><h3>aaa</h3><div><p>bbb<span>ccc</span></p></div></div>',
-            '<h3>aaa</h3>',
-            '<div><p>bbb<span>ccc</span></p></div>']
+    def is_not_p(el):
+        if el.tag == 'p':
+            return False
+        return True
 
-        def is_not_p(el):
-            if el.tag == 'p':
-                return False
-            return True
-
-        doc = fromstring(data)
+    def compare(module):
+        doc = module.fromstring(data)
         output = []
         for el in clean.conditioned_iter(doc, is_not_p):
-            output.append(tostring(el))
+            output.append(module.tostring(el, encoding='unicode'))
         assert list(output) == expected
+
+    compare(lxml.html)
+    compare(lxml_html)